ver. 0.4.0/20170928

+++ hois_back/src/main/java/ee/hitsa/ois/web/ClassifierController.java
@@ -17,12 +17,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import ee.hitsa.ois.domain.Classifier;
-import ee.hitsa.ois.exception.AssertionFailedException;
 import ee.hitsa.ois.service.AutocompleteService;
 import ee.hitsa.ois.service.ClassifierService;
 import ee.hitsa.ois.service.security.HoisUserDetails;
 import ee.hitsa.ois.util.EntityUtil;
 import ee.hitsa.ois.util.StreamUtil;
+import ee.hitsa.ois.util.UserUtil;
 import ee.hitsa.ois.util.WithEntity;
 import ee.hitsa.ois.web.commandobject.ClassifierSearchCommand;
 import ee.hitsa.ois.web.dto.ClassifierDto;
@@ -43,7 +43,7 @@ public class ClassifierController {
      */
     @PostMapping
     public ClassifierDto create(HoisUserDetails user, @Valid @RequestBody Classifier classifier) {
-        AssertionFailedException.throwIf(!user.isMainAdmin(), "Only main administrator can create classifiers");
+        UserUtil.assertIsMainAdmin(user);
         return get(classifierService.save(classifier));
     }
 
@@ -52,7 +52,7 @@ public class ClassifierController {
      */
     @PutMapping("/{code}")
     public ClassifierDto update(HoisUserDetails user, @WithEntity("code") Classifier classifier, @Valid @RequestBody Classifier newClassifier) {
-        AssertionFailedException.throwIf(!user.isMainAdmin(), "Only main administrator can update classifiers");
+        UserUtil.assertIsMainAdmin(user);
         EntityUtil.bindToEntity(newClassifier, classifier);
         return get(classifierService.save(classifier));
     }
@@ -90,7 +90,7 @@ public class ClassifierController {
      */
     @DeleteMapping("/{code}")
     public boolean delete(HoisUserDetails user, @PathVariable("code") String code) {
-        AssertionFailedException.throwIf(!user.isMainAdmin(), "Only main administrator can delete classifiers");
+        UserUtil.assertIsMainAdmin(user);
         classifierService.delete(code);
         return true;
     }
